+++ src/PSRule/Commands/LanguageBlock.cs
@@ -1,6 +1,5 @@
 ﻿using PSRule.Annotations;
 using PSRule.Host;
-using PSRule.Pipeline;
 using PSRule.Rules;
 using System.Collections;
 using System.Management.Automation;