+++ tests/PSRule.Tests/FromFileWithException.Rule.ps1
@@ -1,0 +1,7 @@
+#
+# Pester unit test rules
+#
+
+# Raise an exception
+$notObject = $Null;
+$notObject.ToString();